backend/requirements.txt
django==5.2
djang-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2==2.9.10
tiktoken==0.9.0
pillow==11.2.1


